+++ op-program/client/l2/engineapi/block_processor.go
@@ -52,13 +52,12 @@ func NewBlockProcessorFromPayloadAttributes(provider BlockDataProvider, parent c
 		ParentBeaconRoot: attrs.ParentBeaconBlockRoot,
 	}
 	if attrs.EIP1559Params != nil {
-		// Do we need to check if holocene is active?
 		d, e := eip1559.DecodeHolocene1559Params(attrs.EIP1559Params[:])
 		if d == 0 {
 			d = provider.Config().BaseFeeChangeDenominator(header.Time)
 			e = provider.Config().ElasticityMultiplier()
 		}
-		header.Extra = eip1559.EncodeHoloceneExtraData(uint32(d), uint32(e))
+		header.Extra = eip1559.EncodeHoloceneExtraData(d, e)
 	}
 
 	return NewBlockProcessorFromHeader(provider, header)
